+++ packages/codegen-ui-react/lib/forms/form-renderer-helper/model-fields.ts
@@ -13,7 +13,7 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  */
-import { factory, NodeFlags, ObjectLiteralElementLike } from 'typescript';
+import { factory, NodeFlags, ObjectLiteralElementLike, SyntaxKind } from 'typescript';
 import { FieldConfigMetadata } from '@aws-amplify/codegen-ui';
 
 /**
@@ -38,7 +38,7 @@ export const buildModelFieldObject = (
   const fieldSet = new Set<string>();
   const fields = Object.keys(fieldConfigs).reduce<ObjectLiteralElementLike[]>((acc, value) => {
     const fieldName = value.split('.')[0];
-    const { sanitizedFieldName } = fieldConfigs[value];
+    const { sanitizedFieldName, dataType } = fieldConfigs[value];
     const renderedFieldName = sanitizedFieldName || fieldName;
     if (!fieldSet.has(renderedFieldName)) {
       let assignment: ObjectLiteralElementLike = factory.createShorthandPropertyAssignment(
@@ -53,6 +53,23 @@ export const buildModelFieldObject = (
         assignment = factory.createPropertyAssignment(
           factory.createStringLiteral(fieldName),
           factory.createIdentifier(sanitizedFieldName),
+        );
+      }
+
+      /*
+       Empty string value for not required url field fails to save at datastore
+       let modelFields = {
+        url: url || undefined,
+       }
+      */
+      if (dataType === 'AWSURL' && !shouldBeConst) {
+        assignment = factory.createPropertyAssignment(
+          factory.createStringLiteral(fieldName),
+          factory.createBinaryExpression(
+            factory.createIdentifier(renderedFieldName),
+            factory.createToken(SyntaxKind.BarBarToken),
+            factory.createIdentifier('undefined'),
+          ),
         );
       }