+++ packages/codegen-ui-react/lib/forms/form-renderer-helper/type-helper.ts
@@ -289,7 +289,7 @@ export const validationFunctionType = factory.createTypeAliasDeclaration(
     - onSuccess(fields)
     - onError(fields, errorMessage)
    */
-export const buildFormPropNode = (form: StudioForm) => {
+export const buildFormPropNode = (form: StudioForm, modelName?: string) => {
   const {
     name: formName,
     dataType: { dataSourceType, dataTypeName },
@@ -309,7 +309,7 @@ export const buildFormPropNode = (form: StudioForm) => {
           undefined,
           factory.createIdentifier(lowerCaseFirst(dataTypeName)),
           factory.createToken(SyntaxKind.QuestionToken),
-          factory.createTypeReferenceNode(factory.createIdentifier(dataTypeName), undefined),
+          factory.createTypeReferenceNode(factory.createIdentifier(modelName ?? dataTypeName), undefined),
         ),
       );
     }