+++ packages/codegen-ui-react/lib/imports/import-collection.ts
@@ -15,10 +15,21 @@
  */
 import { ImportDeclaration, factory } from 'typescript';
 import path from 'path';
+import { ComponentMetadata, reservedWords } from '@aws-amplify/codegen-ui';
 import { ImportMapping, ImportValue, ImportSource } from './import-mapping';
+import { isPrimitive } from '../primitive';
+import { createUniqueName } from '../helpers';
 
 export class ImportCollection {
+  constructor(componentMetadata?: ComponentMetadata) {
+    this.importedNames = new Set(Object.values(componentMetadata?.componentNameToTypeMap || {}).concat(reservedWords));
+  }
+
+  importedNames: Set<string>;
+
   #collection: Map<string, Set<string>> = new Map();
+
+  importAlias: Map<string, Map<string, string>> = new Map();
 
   addMappedImport(importValue: ImportValue) {
     const importPackage = ImportMapping[importValue];
@@ -34,11 +45,39 @@ export class ImportCollection {
 
     if (!existingPackage?.has(importName)) {
       existingPackage?.add(importName);
+      if (packageName !== ImportSource.LOCAL_MODELS) {
+        this.importedNames.add(importName);
+      }
     }
+
+    if (packageName === ImportSource.LOCAL_MODELS) {
+      const existingPackageAlias = this.importAlias.get(packageName);
+      const existingAlias = existingPackageAlias?.get(importName);
+      if (existingAlias) return existingAlias;
+      const modelAlias = createUniqueName(
+        importName,
+        (input) =>
+          this.importedNames.has(input) || (input.endsWith('Props') && isPrimitive(input.replace(/Props/g, ''))),
+      );
+      if (existingPackageAlias) {
+        existingPackageAlias.set(importName, modelAlias);
+      } else {
+        const aliasMap = new Map();
+        aliasMap.set(importName, modelAlias);
+        this.importAlias.set(packageName, aliasMap);
+      }
+      this.importedNames.add(modelAlias);
+      return modelAlias;
+    }
+    return importName;
   }
 
   removeImportSource(packageImport: ImportSource) {
     this.#collection.delete(packageImport);
+  }
+
+  getMappedAlias(packageName: string, importName: string) {
+    return this.importAlias.get(packageName)?.get(importName) || importName;
   }
 
   mergeCollections(otherCollection: ImportCollection) {
@@ -87,6 +126,28 @@ export class ImportCollection {
       .concat(
         Array.from(this.#collection).map(([moduleName, imports]) => {
           const namedImports = [...imports].filter((namedImport) => namedImport !== 'default').sort();
+          const aliasMap = this.importAlias.get(moduleName);
+          if (aliasMap) {
+            const importClause = factory.createImportClause(
+              false,
+              undefined,
+              factory.createNamedImports(
+                [...imports].map((item) => {
+                  const alias = aliasMap.get(item);
+                  return factory.createImportSpecifier(
+                    alias && alias !== item ? factory.createIdentifier(item) : undefined,
+                    factory.createIdentifier(alias ?? item),
+                  );
+                }),
+              ),
+            );
+            return factory.createImportDeclaration(
+              undefined,
+              undefined,
+              importClause,
+              factory.createStringLiteral(moduleName),
+            );
+          }
           return factory.createImportDeclaration(
             undefined,
             undefined,