+++ packages/codegen-ui/lib/utils/index.ts
@@ -21,3 +21,4 @@ export * from './form-component-metadata';
 export * from './form-to-component';
 export * from './breakpoint-utils';
 export * from './form-utils';
+export * from './reserved-words';