+++ packages/test-generator/integration-test-templates/src/models/schema.js
@@ -529,4 +529,5 @@ export const schema = {
     },
   },
   version: 'f6252c821249b6b1abda9fb24481c5a4',
+  codegenVersion: '3.2.0',
 };