+++ packages/test-generator/lib/index.ts
@@ -17,3 +17,4 @@ export * from './components';
 export * from './generators';
 export * from './themes';
 export * from './forms';
+export * from './views';