+++ packages/codegen-ui-golden-files/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.4](https://github.com/aws-amplify/amplify-codegen-ui/compare/v2.5.2...v2.5.4) (2022-11-07)
+
+**Note:** Version bump only for package @aws-amplify/codegen-ui-golden-files
+
+## [2.5.3](https://github.com/aws-amplify/amplify-codegen-ui/compare/v2.5.2...v2.5.3) (2022-11-04)
+
+**Note:** Version bump only for package @aws-amplify/codegen-ui-golden-files
+
 ## [2.5.2](https://github.com/aws-amplify/amplify-codegen-ui/compare/v2.4.1...v2.5.2) (2022-10-18)
 
 **Note:** Version bump only for package @aws-amplify/codegen-ui-golden-files